requirements.txt
# Libraries required for the Python project
asyncio==3.4.3  # Asynchronous I/O library for concurrent code execution.
certifi==2022.12.7  # Provides a collection of root certificates for secure SSL/TLS connections.
charset-normalizer==3.1.0  # Charset detection and normalization library for working with character encodings.
fake-useragent==1.1.3  # Generates fake User-Agent strings for web scraping or automation tasks.
idna==3.4  # Library to support the Internationalized Domain Names in Applications (IDNA) protocol.
kafka-python==2.0.2  # Python client for Apache Kafka, a distributed streaming platform.
praw==7.7.0  # Python Reddit API Wrapper for interacting with the Reddit API.
prawcore==2.3.0  # Low-level library for accessing the Reddit API used by praw.
requests==2.28.2  # HTTP library for making requests to web servers.
update-checker==0.18.0  # Library for checking whether a package is up-to-date.
urllib3==1.26.15  # HTTP client library for making HTTP requests with various features.
websocket-client==1.5.1  # WebSocket client implementation for Python to enable WebSocket communication.